refactor: extract data export logic to own ExportModule and ExportService
