+++ src/app/core/export/export-data-button/export-data-button.component.ts
@@ -1,15 +1,15 @@
 import { Component, Input } from "@angular/core";
-import { BackupService } from "../services/backup.service";
+import { ExportService } from "../export-service/export.service";
 
 /**
  * Generic export data button that allows the user to download a file of the given data.
  */
 @Component({
-  selector: "app-export-data",
-  templateUrl: "./export-data.component.html",
-  styleUrls: ["./export-data.component.scss"],
+  selector: "app-export-data-button",
+  templateUrl: "./export-data-button.component.html",
+  styleUrls: ["./export-data-button.component.scss"],
 })
-export class ExportDataComponent {
+export class ExportDataButtonComponent {
   /** data to be exported */
   @Input() data: any = [];
 
@@ -21,7 +21,7 @@ export class ExportDataComponent {
 
   @Input() disabled: boolean = false;
 
-  constructor(private backupService: BackupService) {}
+  constructor(private exportService: ExportService) {}
 
   /**
    * Trigger the download of the export file.
@@ -46,10 +46,10 @@ export class ExportDataComponent {
     let result = "";
     switch (this.format.toLowerCase()) {
       case "json":
-        result = this.backupService.createJson(this.data);
+        result = this.exportService.createJson(this.data);
         return new Blob([result], { type: "application/json" });
       case "csv":
-        result = this.backupService.createCsv(this.data);
+        result = this.exportService.createCsv(this.data);
         return new Blob([result], { type: "text/csv" });
       default:
         console.warn("Not supported format:", this.format);
